feat: move timer to separate component
